Move TestExperiment to testutils package
