+++ src/test/java/uk/ac/ebi/atlas/experimentimport/analyticsindex/AnalyticsIndexerManagerTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.ac.ebi.atlas.model.experiment.ExperimentBuilder.TestExperimentBuilder;
-import uk.ac.ebi.atlas.model.experiment.ExperimentTest.TestExperiment;
+import uk.ac.ebi.atlas.testutils.ExperimentBuilder.TestExperimentBuilder;
+import uk.ac.ebi.atlas.testutils.TestExperiment;
 import uk.ac.ebi.atlas.trader.ExperimentTrader;
 import uk.ac.ebi.atlas.utils.BioentityIdentifiersReader;
 import uk.ac.ebi.atlas.utils.ExperimentSorter;
